Finished tests

+++ Part1/src/main/java/Assignment3/Part1/Compute.java
@@ -3,11 +3,11 @@ package Assignment3.Part1;
 public interface Compute {
 	public int computeTax(int amount);
 	
-	public double weeklySalary(Employee emp);
+	public int weeklySalary(Employee emp);
 	
-	public double fortnightSalary(Employee emp);
+	public int fortnightSalary(Employee emp);
 	
-	public int computeKiwiSaver(Employee emp);
+	public double computeKiwiSaver(Employee emp);
 	
 
 }
